+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   - `onnxruntime` mark all import paths as non-critical
   - `measure` handle current kanvas stage
   - `caption` button in standard-ui
+  - `python` experimental/ignore version checks
 
 ## Update for 2026-06-16